requirements.txt
requests
python-dotenv
google-generativeai
chromadb
sentence-transformers
flask